[AudioFlamingo3] Fix dtype mismatch when loading with bfloat16
In AudioFlamingo3AudioEncoder.forward(), the positional embedding addition
was missing a dtype cast:

    hidden_states = inputs_embeds + self.embed_positions.weight

Since _keep_in_fp32_modules_strict keeps embed_positions in float32,
this addition upcasts hidden_states to float32. The subsequent LayerNorm,
whose weight is bfloat16, then raises:
  RuntimeError: expected scalar type Float but found BFloat16

Fix: cast the result back to inputs_embeds.dtype, matching the pattern
already used in VoxtralEncoder.forward:
    hidden_states = (inputs_embeds + self.embed_positions.weight).to(inputs_embeds.dtype)

Verified: AudioFlamingo3 on MMAU-full (9000 samples), bf16 + flash_attn2,
bs=20 → 72.49% accuracy (matches bs=1 baseline of 72.44%).

+++ src/transformers/models/audioflamingo3/modeling_audioflamingo3.py
@@ -382,7 +382,7 @@ class AudioFlamingo3Encoder(AudioFlamingo3PreTrainedModel):
         inputs_embeds = inputs_embeds.permute(0, 2, 1)
 
         # Add positions, dropout
-        hidden_states = inputs_embeds + self.embed_positions.weight
+        hidden_states = (inputs_embeds + self.embed_positions.weight).to(inputs_embeds.dtype)
         hidden_states = nn.functional.dropout(hidden_states, p=self.dropout, training=self.training)
 
         attention_mask = create_bidirectional_mask(
